+++ low_poly_engine/LPEnginePrimitives.m
@@ -27,7 +27,7 @@ static float pointSize = 10;
 static int16_t *depthBuffer;
 static NSInteger vertexMemoryAllocationSize=0;
 
-LPPoint LPPointMake(int16_t x, int16_t y, int16_t z) {
+LPPoint LPPointMake(float x, float y, float z) {
     LPPoint newPoint;
     newPoint.x = x;
     newPoint.y = y;
